+++ core/docs/changelog.md
@@ -1,5 +1,15 @@
 # TBD
 
+# 1.15.3
+### Features
+* Add custom params log to show users that Kurtosis has loaded this configuration
+
+### Changes
+* Change the API container to depend on `kurtosis-client`, rather than duplicating the `.proto` file in here
+* Change the API container to depend on `kurtosis-libs`, rather than duplicating the testsuite `.proto` file in here
+* Removed the `regenerate-protobuf-bindings.sh`, as it's no longer necessary
+* Upgraded Kurtosis Client version to 0.2.0 (was 0.1.1)
+
 # 1.15.2
 ### Fixes
 * Fixed an issue with the LogStreamer where StdCopy is a blocking method which could halt all of Kurtosis if not dealt with appropriately